Revert "Refactor bjorn code into separate file. (#309)"
This reverts commit a9a4d0fc229f1c7a9616635a6903e9dedf3a971c.
